refactor: extract HasHttpMethod concern to dedupe method() coercion
The identical HttpMethod-coercion method() lived in Action, Link, MenuItem
and Form. Pull it into a HasHttpMethod concern alongside HasVariant.
